+++ arangod/RestServer/TtlFeature.h
@@ -24,14 +24,14 @@
 #pragma once
 
 #include "ApplicationFeatures/ApplicationFeature.h"
-#include "RestServer/TtlFeatureOptions.h"
+#include "RestServer/TtlProperties.h"
 
 #include <mutex>
 
 namespace arangodb {
 namespace velocypack {
-class Builder;
-class Slice;
+    class Builder;
+    class Slice;
 }  // namespace velocypack
 
 class TtlThread;
@@ -55,16 +55,6 @@ struct TtlStatistics {
   TtlStatistics& operator+=(arangodb::velocypack::Slice const& other);
 
   void toVelocyPack(arangodb::velocypack::Builder& out) const;
-};
-
-struct TtlProperties {
-  static constexpr uint64_t minFrequency = 1 * 1000;  // milliseconds
-  uint64_t frequency = 30 * 1000;                     // milliseconds
-  uint64_t maxTotalRemoves = 1000000;
-  uint64_t maxCollectionRemoves = 100000;
-
-  void toVelocyPack(arangodb::velocypack::Builder& out, bool isActive) const;
-  Result fromVelocyPack(arangodb::velocypack::Slice const& properties);
 };
 
 class TtlFeature final : public application_features::ApplicationFeature {
@@ -105,9 +95,6 @@ class TtlFeature final : public application_features::ApplicationFeature {
 
  private:
   void shutdownThread() noexcept;
-
- private:
-  TtlFeatureOptions _options;
 
   /// @brief protects _properties and _active
   mutable std::mutex _propertiesMutex;
