+++ arangod/arangoserver.cmake
@@ -184,6 +184,7 @@ add_library(arangoserver STATIC
   RestServer/TimeZoneFeature.cpp
   RestServer/TtlFeature.cpp
   RestServer/TtlOptionsProvider.cpp
+  RestServer/TtlProperties.cpp
   RestServer/UpgradeFeature.cpp
   RestServer/UpgradeOptionsProvider.cpp
   VectorIndex/VectorIndexFeature.cpp
